refactor(tasks): replace TaskProgressModal with TaskDetail navigation
- Delete TaskProgressModal; watchTask now navigates to /tasks/:id
- Add /tasks/:id route pointing to TaskDetail (was an orphan page)
- Extract shared step-parsing into useTaskSteps composable
- Remove duplicated stepSections computed from Tasks.vue and TaskDetail.vue
