+++ admin/frontend/src/pages/Tasks.vue
@@ -3,6 +3,7 @@ import { ref, computed, watch, onMounted, onUnmounted } from 'vue'
 import { useRoute, useRouter } from 'vue-router'
 import { Button, Badge, Dialog, FormControl, LoadingText, ErrorMessage } from 'frappe-ui'
 import TerminalOutput from '../components/TerminalOutput.vue'
+import { useTaskSteps } from '../composables/useTaskSteps.js'
 import { processLine } from '../utils/ansi.js'
 import LucideCheck from '~icons/lucide/check'
 import LucideLoader2 from '~icons/lucide/loader-2'
@@ -103,31 +104,7 @@ watch(selectedTaskId, (id) => {
   if (id) loadDetail(id)
 })
 
-const stepSections = computed(() => {
-  const markers = []
-  rawLines.value.forEach((line, idx) => {
-    const m = line.match(/^##\[step:(\w+),([\d.]+)\]\s*(.*)/)
-    if (m) markers.push({ key: m[1], ts: parseFloat(m[2]) * 1000, label: m[3].trim(), idx })
-  })
-  const sections = []
-  for (let i = 0; i < markers.length; i++) {
-    const m = markers[i]
-    if (m.key === 'done') break
-    const next = markers[i + 1]
-    const lineStart = m.idx + 1
-    const lineEnd = next ? next.idx : rawLines.value.length
-    const endedAt = next ? next.ts : null
-    let status
-    if (next) status = 'done'
-    else if (!streaming.value && task.value?.status === 'failed') status = 'failed'
-    else if (!streaming.value) status = 'done'
-    else status = 'running'
-    sections.push({ key: m.key, label: m.label, startedAt: m.ts, endedAt, lineStart, lineEnd, status })
-  }
-  return sections
-})
-
-const hasSteps = computed(() => stepSections.value.length > 0)
+const { stepSections, hasSteps, stepDuration } = useTaskSteps(rawLines, streaming, task)
 
 function sectionLines(section) {
   return rawLines.value
@@ -140,12 +117,6 @@ function sectionHasOutput(section) {
   return rawLines.value
     .slice(section.lineStart, section.lineEnd)
     .some(l => l.trim() && !l.match(/^##\[step:/))
-}
-
-function stepDuration(section) {
-  if (!section.startedAt || !section.endedAt) return null
-  const s = (section.endedAt - section.startedAt) / 1000
-  return s < 60 ? `${s.toFixed(1)}s` : `${(s / 60).toFixed(1)}m`
 }
 
 function toggleStep(key) {
