+++ admin/frontend/src/router.js
@@ -7,7 +7,7 @@ const routes = [
   { path: '/monitor', component: () => import('./pages/Monitor.vue'), meta: { title: 'Monitor' } },
   { path: '/logs', component: () => import('./pages/Logs.vue'), meta: { title: 'Logs' } },
   { path: '/tasks', component: () => import('./pages/Tasks.vue'), meta: { title: 'Tasks' } },
-  { path: '/tasks/:id', redirect: to => ({ path: '/tasks', query: { task: to.params.id } }) },
+  { path: '/tasks/:id', component: () => import('./pages/TaskDetail.vue'), meta: { title: 'Tasks' } },
   { path: '/database', component: () => import('./pages/Database.vue'), meta: { title: 'Database' } },
   { path: '/database/binlogs/:name', component: () => import('./pages/BinlogDetail.vue'), meta: { title: 'Binlogs' } },
   { path: '/snapshots', component: () => import('./pages/Snapshots.vue'), meta: { title: 'Snapshots' } },
